Clean DnDTable and ParameterTable
