--- src/components/composite/dnd-table-v2/dnd-table.type.ts
@@ -1,99 +1,0 @@
-/**
- * Copyright (c) 2025, RTE (http://www.rte-france.com)
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-
-import { JSX, ReactNode } from 'react';
-import { SxProps, Theme } from '@mui/material';
-import { ElementType } from '../../../utils/types/elementType';
-import { EquipmentType } from '../../../utils/types/equipmentType';
-import { Option } from '../../../utils/types/types';
-
-export const SELECTED = 'selected';
-export const MAX_ROWS_NUMBER = 100;
-
-export enum DndColumnType {
-    TEXT = 'TEXT',
-    NUMERIC = 'NUMERIC',
-    AUTOCOMPLETE = 'AUTOCOMPLETE',
-    SELECT = 'SELECT',
-    CHIP_ITEMS = 'CHIP_ITEMS',
-    DIRECTORY_ITEMS = 'DIRECTORY_ITEMS',
-    SWITCH = 'SWITCH',
-    DESCRIPTIONS = 'DESCRIPTIONS',
-    CUSTOM = 'CUSTOM',
-}
-
-export interface ColumnBase {
-    dataKey: string;
-    maxWidth?: number | string;
-    width?: number | string;
-    label?: string;
-    sxHeader?: SxProps<Theme>;
-    extra?: JSX.Element;
-    editable?: boolean;
-    type: DndColumnType;
-    initialValue?: any; // should conform to the type field
-    // to force to propagate a change event to the parent component via callback
-    // for example, refresh contingency count against a change of selected contingency
-    shouldHandleOnChangeCell?: boolean;
-    hideErrorMessage?: boolean;
-}
-
-export interface ColumnText extends ColumnBase {
-    type: DndColumnType.TEXT;
-}
-
-export interface ColumnNumeric extends ColumnBase {
-    type: DndColumnType.NUMERIC;
-    adornment?: { text: string };
-    clearable?: boolean;
-    textAlign?: 'right' | 'left';
-}
-
-export interface ColumnAutocomplete extends ColumnBase {
-    type: DndColumnType.AUTOCOMPLETE;
-    options: Option[];
-}
-
-export interface ColumnSelect extends ColumnBase {
-    type: DndColumnType.SELECT;
-    options: Option[];
-}
-
-export interface ColumnDirectoryItem extends ColumnBase {
-    type: DndColumnType.DIRECTORY_ITEMS;
-    equipmentTypes: EquipmentType[];
-    elementType: ElementType;
-    titleId: string;
-}
-
-export interface ColumnChipsItem extends ColumnBase {
-    type: DndColumnType.CHIP_ITEMS;
-}
-
-export interface ColumnSwitchItem extends ColumnBase {
-    type: DndColumnType.SWITCH;
-}
-
-export interface ColumnDescriptions extends ColumnBase {
-    type: DndColumnType.DESCRIPTIONS;
-}
-
-export interface ColumnCustom extends ColumnBase {
-    type: DndColumnType.CUSTOM;
-    component: (rowIndex: number) => ReactNode;
-}
-
-export type DndColumn =
-    | ColumnNumeric
-    | ColumnAutocomplete
-    | ColumnSelect
-    | ColumnText
-    | ColumnDirectoryItem
-    | ColumnChipsItem
-    | ColumnSwitchItem
-    | ColumnDescriptions
-    | ColumnCustom;
