--- src/components/composite/dnd-table-v2/index.ts
@@ -1,12 +1,0 @@
-/**
- * Copyright (c) 2025, RTE (http://www.rte-france.com)
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-export * from './dnd-table';
-export * from './dnd-table.type';
-export * from './dnd-table-utils';
-export * from './dnd-table-add-rows-dialog';
-export * from './dnd-table-bottom-left-buttons';
-export * from './dnd-table-bottom-right-buttons';
