+++ src/components/composite/dnd-table/deletable-table-row.tsx
