+++ src/components/composite/dnd-table/dnd-table-utils.ts
