--- src/features/parameters/common/parameter-table/constants.ts
@@ -1,13 +1,0 @@
-/**
- * Copyright (c) 2026, RTE (http://www.rte-france.com)
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-
-export const MAX_ROWS_NUMBER = 100;
-
-export const NAME = 'name';
-export const ID = 'id';
-export const ACTIVATED = 'activated';
-export const DESCRIPTION = 'description';
